requirements.txt
Django==1.11
django-oauth-toolkit==0.12.0
djangorestframework==3.6.2

